backend/app/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==43.0.1
cssselect==1.2.0
dataclasses-json==0.6.7
deepdiff==8.0.1
dnspython==2.6.1
docopt==0.6.2
emoji==2.13.2
eutils==0.6.0
exceptiongroup==1.2.2
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.62.3
gunicorn
h11==0.14.0
habanero==1.2.6
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-google-genai==2.0.0
langchain-mongodb==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.129
Levenshtein==0.26.0
lxml==5.3.0
lxml_html_clean==0.2.2
MarkupSafe==2.1.5
marshmallow==3.22.0
metapub==0.5.12
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.9.1
pyparsing==3.1.4
pypdf==5.0.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-Levenshtein==0.26.0
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tabulate==0.9.0
tenacity==8.5.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
Unidecode==1.3.8
unstructured==0.15.13
unstructured-client==0.25.9
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.13.1
yt-dlp==2024.9.27
zipp==3.20.2
